Refactor loaders
