+++ docbrown/src/db/mod.rs
@@ -1,4 +1,3 @@
-pub mod csv_loader;
 pub mod edge;
 pub mod graph;
 pub mod graph_immutable;
@@ -6,7 +5,6 @@ pub mod graph_layer;
 pub mod graph_window;
 pub mod path;
 pub mod perspective;
-pub mod polars_loader;
 pub mod program;
 pub mod vertex;
 pub mod vertices;
