+++ docbrown/src/graph_loader/example/mod.rs
@@ -1,0 +1,3 @@
+pub mod lotr_graph;
+pub mod reddit_hyperlinks;
+pub mod sx_superuser_graph;
