+++ examples/rust/src/bin/healthcheck/main.rs
@@ -9,7 +9,7 @@ mod test {
 
     use docbrown::algorithms::connected_components::weakly_connected_components;
     use docbrown::core::Direction;
-    use docbrown::db::csv_loader::CsvLoader;
+    use docbrown::graph_loader::source::csv_loader::CsvLoader;
     use docbrown::db::{
         graph::Graph,
         view_api::*,
