+++ examples/rust/src/bin/hulongbay/main.rs
@@ -14,7 +14,7 @@ use docbrown::algorithms::triangle_count::triangle_counting_fast;
 use docbrown::core::tgraph::TemporalGraph;
 use docbrown::core::{state, utils};
 use docbrown::core::{Direction, Prop};
-use docbrown::db::csv_loader::CsvLoader;
+use docbrown::graph_loader::souce::csv_loader::CsvLoader;
 use docbrown::db::graph::Graph;
 use docbrown::db::program::{GlobalEvalState, Program};
 use docbrown::db::view_api::*;
